Move pending transactions caching logic to compat layer
