+++ services/core/shared/core/compat/sdk_v2/index.js
@@ -95,6 +95,7 @@ module.exports = {
 	calculateWeightedAvg: nop,
 	updateFinalizedHeight,
 	getPendingTransactions,
+	loadAllPendingTransactions: nop,
 	events,
 	init: nop,
 };
